feat(setup): simplify local setup to bun install && bun dev (#885)
- bun dev now bootstraps .env.local (generated secrets) via scripts/ensure-env.ts
- setup.ts cut from 2,290 to 121 lines (local AI-key prompts only); prod/deploy/
  pr-preview flows moved to scripts/setup-prod.ts minus dead QStash/Docker/preview
  sections, invoked via bun setup --prod|--deploy|--pr-preview
- package.json scripts pruned from 64 to 39 entries; dev absorbs the Stripe listener
- Deploy to Cloudflare button in README + docs, with .dev.vars.example for the
  button's secret prompts; default-env tail_consumers emptied so button deploys
  and local dev don't reference the prod log forwarder
- Remove dead env vars everywhere: FAL/LLM_CONCURRENCY_LIMIT, CEREBRAS_API_KEY,
  LETZAI_API_KEY, LANGFUSE_TRACING_ENVIRONMENT, VITE_CONTACT/PRIVACY_EMAIL
  (now derived from VITE_APP_URL), UPSTASH_REDIS_REST_* in CI workflows
- De-stale README/CONTRIBUTING/getting-started/deployment docs (QStash, Docker,
  SQLite, bun:test references)

Closes #885
