+++ .env.example
@@ -1,19 +1,19 @@
 # =============================================================================
 # Environment Configuration
-# Recommended: run `bun setup` for interactive configuration
+# `bun dev` auto-generates a working .env.local on first run — you only need
+# this file as a reference for optional services. `bun setup` adds AI keys.
 # =============================================================================
 
 # =============================================================================
-# REQUIRED - App won't start without these
+# REQUIRED - auto-generated by `bun dev` if missing
 # =============================================================================
 
-# App Configuration
+# App Configuration (contact/privacy emails shown on marketing pages are
+# derived from VITE_APP_URL's domain — hello@<domain> / privacy@<domain>)
 VITE_APP_URL=http://localhost:3000
 VITE_APP_NAME=OpenStory
-VITE_CONTACT_EMAIL=hello@example.com
-VITE_PRIVACY_EMAIL=privacy@example.com
 
-# Auth (auto-generated by `bun setup`)
+# Auth (auto-generated)
 BETTER_AUTH_SECRET=
 
 # Database - Cloudflare D1
@@ -33,25 +33,27 @@ BETTER_AUTH_SECRET=
 
 # FAL_KEY=           # https://fal.ai - image & video generation
 # OPENROUTER_KEY=    # https://openrouter.ai - LLM access for script analysis
-# CEREBRAS_API_KEY=  # https://cerebras.ai - fast inference (optional)
-# LETZAI_API_KEY=    # https://letz.ai - alternative image gen (optional)
 
 # =============================================================================
-# STORAGE - S3-compatible object storage for generated media
+# STORAGE - Cloudflare R2 (native bindings in wrangler.jsonc)
 #
 # Local dev needs NONE of these: storage uses the local Miniflare R2 binding
 # and the worker's /r2/$ route serves reads when R2_PUBLIC_STORAGE_DOMAIN is
-# unset. Set them only to opt into remote R2 (also requires "remote": true on
-# the binding in wrangler.jsonc).
+# unset. To opt into remote R2, set R2_PUBLIC_STORAGE_DOMAIN here and
+# "remote": true on the binding in wrangler.jsonc.
 # =============================================================================
 
+# R2_PUBLIC_STORAGE_DOMAIN=storage.example.com
+# R2_PUBLIC_ASSETS_BUCKET=my-app-assets              # Shared across all environments
+# VITE_R2_PUBLIC_ASSETS_DOMAIN=assets.example.com     # Same for staging and production
+
+# S3 API credentials — only needed by the content upload scripts
+# (scripts/upload-style-previews-to-r2.ts, upload-style-sample-videos-to-r2.ts).
+# The app itself never uses them.
 # R2_ACCOUNT_ID=
 # R2_ACCESS_KEY_ID=
 # R2_SECRET_ACCESS_KEY=
 # R2_BUCKET_NAME=my-app-storage
-# R2_PUBLIC_STORAGE_DOMAIN=storage.example.com
-# R2_PUBLIC_ASSETS_BUCKET=my-app-assets              # Shared across all environments
-# VITE_R2_PUBLIC_ASSETS_DOMAIN=assets.example.com     # Same for staging and production
 
 # Realtime progress updates run on a Cloudflare Durable Object (`REALTIME`
 # binding in wrangler.jsonc) — no external Redis/Upstash config required.
@@ -76,7 +78,7 @@ BETTER_AUTH_SECRET=
 # STRIPE_WEBHOOK_SECRET=
 
 # BYOK API Key Encryption (required if users bring their own AI keys)
-# Generate with: openssl rand -base64 32
+# Auto-generated by `bun dev` in .env.local
 # API_KEY_ENCRYPTION_KEY=
 
 # Observability - LLM tracing via Langfuse (optional, gracefully degrades)
@@ -91,7 +93,3 @@ BETTER_AUTH_SECRET=
 
 # System admin emails (comma-separated, for gift code management)
 # ADMIN_EMAILS=admin@example.com
-
-# AI generation concurrency
-# FAL_CONCURRENCY_LIMIT=10
-# LLM_CONCURRENCY_LIMIT=5
