+++ .github/workflows/deploy-cloudflare.yml
@@ -233,24 +233,16 @@ jobs:
           SECRET_VITE_APP_NAME: ${{ vars.APP_NAME }}
           SECRET_VITE_APP_URL: ${{ steps.deploy.outputs.url }}
           SECRET_BETTER_AUTH_SECRET: ${{ secrets.BETTER_AUTH_SECRET }}
-          SECRET_CEREBRAS_API_KEY: ${{ secrets.CEREBRAS_API_KEY }}
           SECRET_EMAIL_FROM: ${{ secrets.EMAIL_FROM }}
-          SECRET_FAL_CONCURRENCY_LIMIT: ${{ secrets.FAL_CONCURRENCY_LIMIT }}
           SECRET_FAL_KEY: ${{ secrets.FAL_KEY }}
           SECRET_GOOGLE_CLIENT_ID: ${{ secrets.GOOGLE_CLIENT_ID }}
           SECRET_GOOGLE_CLIENT_SECRET: ${{ secrets.GOOGLE_CLIENT_SECRET }}
           SECRET_LANGFUSE_BASE_URL: ${{ secrets.LANGFUSE_BASE_URL }}
           SECRET_LANGFUSE_PUBLIC_KEY: ${{ secrets.LANGFUSE_PUBLIC_KEY }}
           SECRET_LANGFUSE_SECRET_KEY: ${{ secrets.LANGFUSE_SECRET_KEY }}
-          SECRET_LANGFUSE_TRACING_ENVIRONMENT: ${{ secrets.LANGFUSE_TRACING_ENVIRONMENT }}
-          SECRET_LETZAI_API_KEY: ${{ secrets.LETZAI_API_KEY }}
           SECRET_OPENROUTER_KEY: ${{ secrets.OPENROUTER_KEY }}
-          SECRET_UPSTASH_REDIS_REST_TOKEN: ${{ secrets.UPSTASH_REDIS_REST_TOKEN }}
-          SECRET_VITE_CONTACT_EMAIL: ${{ vars.VITE_CONTACT_EMAIL }}
-          SECRET_VITE_PRIVACY_EMAIL: ${{ vars.VITE_PRIVACY_EMAIL }}
           SECRET_VITE_R2_PUBLIC_ASSETS_DOMAIN: ${{ vars.VITE_R2_PUBLIC_ASSETS_DOMAIN }}
           SECRET_R2_PUBLIC_STORAGE_DOMAIN: ${{ vars.R2_PUBLIC_STORAGE_DOMAIN }}
-          SECRET_UPSTASH_REDIS_REST_URL: ${{ secrets.UPSTASH_REDIS_REST_URL }}
           SECRET_VITE_PUBLIC_POSTHOG_PROJECT_TOKEN: ${{ secrets.VITE_PUBLIC_POSTHOG_PROJECT_TOKEN }}
           SECRET_VITE_PUBLIC_POSTHOG_HOST: ${{ vars.VITE_PUBLIC_POSTHOG_HOST }}
         run: |
