+++ CLAUDE.md
@@ -6,7 +6,7 @@ AI-powered video sequence platform built with TanStack Start, deployed to Cloudf
 
 ```bash
 # Dev
-bun dev                            # All-in-one: DB migrate + seed, Vite (Workerd via cf-plugin)
+bun dev                            # All-in-one: env bootstrap, DB migrate + seed, Vite (Workerd via cf-plugin), Stripe listener
 bun storybook                      # Storybook on :6006
 bun db:studio:local                # Inspect local D1 tables (wrangler d1 execute)
 
@@ -44,7 +44,7 @@ bun cf:dev                         # wrangler dev against built worker (preview)
 bun cf:deploy:prd                  # Cloudflare Workers production deploy
 ```
 
-`bun dev` runs vite dev (cf-plugin → Workerd via Miniflare, port 3000) alongside the Stripe listener. The app runs in **Workerd locally** — same runtime as production — so D1, R2 bindings, **Cloudflare Workflows**, env.\* access, and request lifecycle all match prod. No QStash/Docker needed: workflows execute in-process in Workerd.
+`bun dev` runs vite dev (cf-plugin → Workerd via Miniflare, port 3000) alongside the Stripe listener (skipped without `STRIPE_SECRET_KEY`). Its first step (`scripts/ensure-env.ts`) creates `.env.local` with generated secrets if missing, so a fresh clone needs only `bun install && bun dev`. The app runs in **Workerd locally** — same runtime as production — so D1, R2 bindings, **Cloudflare Workflows**, env.\* access, and request lifecycle all match prod. No QStash/Docker needed: workflows execute in-process in Workerd.
 
 **Bun-as-launcher pattern:** `bun script.ts` (no `--bun`) keeps Bun as the CLI launcher but executes under **Node**, while still autoloading `.env*`. Use `bun --env-file=<path>` to override the default `.env.local`. No `--bun` flag should appear in package.json scripts.
 
@@ -93,8 +93,9 @@ teams
 
 ```bash
 bun install
-bun setup                          # Auto-configure local dev (SQLite + secrets)
-bun db:setup                       # Migrate + seed database
+bun dev                            # That's it — env, migrations, seed all happen on first run
+bun setup                          # Optional: add FAL_KEY / OPENROUTER_KEY interactively
+bun setup --prod                   # Production config + deploy (--deploy, --pr-preview also available)
 ```
 
 **Branch + commit conventions:** Branches must be named `<issue-number>-feature-name` (e.g. `393-improve-readme`). Lefthook extracts the issue number and tags commits with `#<issue>` automatically. See `CONTRIBUTING.md`. Lefthook also runs quality checks pre-commit.
