+++ docs/deployment/cloudflare.md
@@ -7,6 +7,16 @@ order: 10
 
 OpenStory deploys to Cloudflare Workers, using D1 (SQLite) for the database and R2 for media storage.
 
+## One-Click Deploy
+
+[![Deploy to Cloudflare](https://deploy.workers.cloudflare.com/button)](https://deploy.workers.cloudflare.com/?url=https://github.com/openstory-so/openstory)
+
+The deploy button clones the repo into your Cloudflare account, provisions the resources declared in `wrangler.jsonc`, prompts for the secrets listed in `.dev.vars.example`, and sets up CI for your copy.
+
+## Guided Setup
+
+From your own clone, `bun setup --prod` walks through everything interactively: production env vars (`.env.production`), R2 domains + CORS, optional services, pushing secrets to Cloudflare and GitHub, and the first deploy. `bun setup --deploy` re-runs just the secrets-push + deploy phase, and `bun setup --pr-preview` pushes preview secrets to the GitHub `staging` environment used by PR preview deploys.
+
 ## Prerequisites
 
 - A [Cloudflare](https://cloudflare.com) account
@@ -29,10 +39,7 @@ The Worker entry point is `src/server.ts` with `nodejs_compat` enabled.
 # Generate Worker types from wrangler.jsonc
 bun cf:typegen
 
-# Build for Cloudflare (sets BUILD_CLOUDFLARE=1 so Vite uses the Cloudflare preset)
-bun cf:build
-
-# Deploy to production (runs cf:build then `wrangler deploy`)
+# Deploy to production (typegen, CLOUDFLARE_ENV=production build, wrangler deploy)
 bun cf:deploy:prd
 ```
 
@@ -56,9 +63,7 @@ Secrets are pushed to the Worker via `wrangler secret bulk`. The full list is de
 | `BETTER_AUTH_SECRET`                        | Better Auth signing secret                                                            |
 | `VITE_APP_URL`                              | Public URL of the deployment                                                          |
 | `FAL_KEY`                                   | fal.ai API key for image/video generation                                             |
-| `QSTASH_TOKEN`                              | QStash token for workflow execution                                                   |
-| `QSTASH_CURRENT_SIGNING_KEY`                | QStash request verification                                                           |
-| `QSTASH_NEXT_SIGNING_KEY`                   | QStash request verification (rotation)                                                |
+| `OPENROUTER_KEY`                            | OpenRouter API key for LLM script analysis                                            |
 | `GOOGLE_CLIENT_ID` / `GOOGLE_CLIENT_SECRET` | Google OAuth credentials                                                              |
 | `EMAIL_FROM`                                | Sender address for transactional email (domain onboarded in Cloudflare Email Service) |
