+++ drizzle.config.d1.ts
@@ -5,7 +5,6 @@ import { defineConfig } from 'drizzle-kit';
  *
  * D1-specific commands:
  *   bun db:migrate:d1   # Apply migrations to remote D1 via HTTP API
- *   bun db:push:d1      # Push schema directly to remote D1
  *   bun db:studio:d1    # Open Drizzle Studio connected to D1
  *
  * Requires env vars: CLOUDFLARE_ACCOUNT_ID, CLOUDFLARE_D1_DATABASE_ID, CLOUDFLARE_API_TOKEN
