+++ package.json
@@ -28,91 +28,44 @@
   },
   "scripts": {
     "setup": "bun scripts/setup.ts",
-    "setup:prd": "bun scripts/setup.ts --prod",
-    "setup:deploy": "bun scripts/setup.ts --deploy",
-    "setup:stg": "bun scripts/setup.ts --pr-preview",
-    "r2:cors": "bun scripts/set-r2-cors.ts",
-    "r2:cors:prd": "bun scripts/set-r2-cors.ts --env=prd",
-    "r2:cors:stg": "bun scripts/set-r2-cors.ts --env=stg",
-    "r2:cors:dev": "bun scripts/set-r2-cors.ts --env=dev",
-    "setup:previews": "bun scripts/generate-style-previews.ts && bun scripts/upload-style-previews-to-r2.ts",
-    "setup:system-previews": "bun scripts/generate-system-previews.ts && bun scripts/upload-system-previews-to-r2.ts",
-    "styles:sample-videos": "bun scripts/generate-style-sample-videos.ts",
-    "styles:sample-videos:score": "bun scripts/score-style-scripts.ts",
-    "styles:sample-videos:upload": "bun scripts/upload-style-sample-videos-to-r2.ts",
-    "styles:sample-videos:seed:local": "bun scripts/seed-style-sample-videos.ts --local",
-    "styles:sample-videos:seed:d1": "bun scripts/seed-style-sample-videos.ts --d1",
-    "upload:prompts": "bun scripts/upload-all-prompts.ts",
-    "agent-harness": "bunx github:tombeckenham/agent-harness",
-    "dev": "bun --sequential db:migrate:local db:seed:local dev:vite",
-    "dev:all": "bun run --parallel dev stripe:dev",
+    "dev": "bun scripts/ensure-env.ts && bun run --parallel dev:app stripe:dev",
+    "dev:app": "bun --sequential db:migrate:local db:seed:local dev:vite",
     "dev:vite": "bun scripts/build-content-collections.ts && vite dev",
     "stripe:dev": "source .env.local 2>/dev/null; [ -z \"${STRIPE_SECRET_KEY}\" ] && echo 'Skipping Stripe (no STRIPE_SECRET_KEY)' && exit 0; stripe listen --forward-to localhost:3000/api/billing/webhook --api-key=${STRIPE_SECRET_KEY}",
     "build": "vite build",
     "build:e2e": "CLOUDFLARE_ENV=test VITE_APP_URL=http://localhost:3001 VITE_DISABLE_DEVTOOLS=true vite build",
-    "preview": "VITE_APP_URL=http://localhost:8787 vite build && VITE_APP_URL=http://localhost:8787 wrangler dev",
     "lint": "bunx oxlint --type-aware",
     "lint:fix": "bunx oxlint --type-aware --fix",
     "dead-code": "bun knip",
-    "dead-code:fix": "bun knip --fix",
     "format": "oxfmt",
     "format:check": "oxfmt --check",
-    "lint:open": "(bun lint 2>&1; bun typecheck 2>&1) | grep -oE '[^ (\\[]+\\.tsx?' | sort -u | sed 's/\\$/\\\\$/g' | tr '\\n' '\\0' | xargs -0 cursor",
     "typecheck": "tsgo --noEmit",
     "test": "vitest run",
-    "test:ui": "vitest",
     "test:coverage": "vitest run --coverage",
     "test:watch": "vitest",
     "test:e2e": "bun playwright test",
     "test:e2e:ui": "bun playwright test --ui",
-    "test:e2e:headed": "bun playwright test --headed",
     "test:e2e:full": "PLAYWRIGHT_FULL_PIPELINE=true bun playwright test full-sequence",
     "test:e2e:full:built": "PLAYWRIGHT_FULL_PIPELINE=true bun run build:e2e && bun run prepare:e2e-built && E2E_BUILT=true PLAYWRIGHT_FULL_PIPELINE=true bun playwright test full-sequence",
     "test:e2e:record": "E2E_RECORD=1 bun playwright test",
     "test:e2e:full:record": "E2E_RECORD=1 PLAYWRIGHT_FULL_PIPELINE=true bun playwright test full-sequence",
     "prepare:e2e-built": "mkdir -p .wrangler/deploy && printf '%s' '{\"configPath\":\"../../dist/server/wrangler.json\",\"auxiliaryWorkers\":[]}' > .wrangler/deploy/config.json",
     "test:e2e:setup": "bun run db:migrate:test && bun scripts/seed.ts --test",
-    "auth:generate": "bunx @better-auth/cli generate --config src/lib/auth/cli-config.ts",
     "db:generate": "bun drizzle-kit generate --config=drizzle.config.d1.ts",
     "db:migrate:local": "bun scripts/migrate-local-d1.ts",
     "db:migrate:test": "bun scripts/migrate-local-d1.ts --test",
     "db:migrate:d1": "bun drizzle-kit migrate --config=drizzle.config.d1.ts",
-    "db:studio:test": "CLOUDFLARE_ENV=test bun db:studio:local",
     "db:studio:local": "bun drizzle-kit studio --config=drizzle.config.local.ts",
     "db:studio:d1": "bun drizzle-kit studio --config=drizzle.config.d1.ts",
     "db:seed:local": "bun scripts/seed.ts --local",
-    "db:seed:test": "bun scripts/seed.ts --test",
     "db:seed:d1": "bun scripts/seed.ts --d1",
-    "db:check": "bun drizzle-kit check --config=drizzle.config.d1.ts",
-    "db:push:d1": "bun drizzle-kit push --config=drizzle.config.d1.ts",
-    "db:check-migrations": "bun scripts/check-migrations.ts",
-    "db:check-migrations:all": "bun scripts/check-migrations.ts --all",
-    "db:setup:local": "bun run db:migrate:local && bun run db:seed:local",
-    "db:fork:local": "bun scripts/db-worktree.ts --fork",
-    "db:promote:local": "bun scripts/db-worktree.ts --promote",
-    "r2:import:local": "bun scripts/import-remote-r2.ts",
-    "memory:profile": "bun scripts/memory-profile.ts",
-    "memory:watch": "bun scripts/memory-profile.ts --watch",
-    "memory:clear": "bun scripts/memory-profile.ts --clear",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build",
     "motion:codegen": "bun scripts/pull-motion-schemas.ts",
     "models:check": "bun scripts/check-model-updates.ts",
-    "doppler:setup": "doppler setup -p openstory -c dev",
-    "doppler:pull:dev": "doppler secrets download -p openstory --config=dev --format=env --no-file > .env.local",
-    "doppler:pull:dev_test": "doppler secrets download -p openstory --config=dev_test --format=env --no-file > .env.local",
-    "doppler:pull:prd": "doppler secrets download -p openstory --config=prd --format=env --no-file > .env.production",
-    "doppler:pull:stg": "doppler secrets download -p openstory --config=stg --format=env --no-file > .env.staging",
-    "doppler:push:dev": "doppler secrets upload -p openstory --config=dev .env.local",
-    "doppler:push:prd": "doppler secrets upload -p openstory --config=prd .env.production",
-    "doppler:push:stg": "doppler secrets upload -p openstory --config=stg .env.staging",
     "cf:typegen": "wrangler types",
     "cf:dev": "wrangler dev",
-    "workflows:explorer": "open http://localhost:${PORT:-3000}/cdn-cgi/explorer",
-    "cf:deploy:prd": "bun run cf:typegen && CLOUDFLARE_ENV=production bun run build && wrangler deploy --env=production",
-    "cf:deploy:forwarder:prd": "wrangler deploy --env prd --config workers/posthog-log-forwarder/wrangler.jsonc",
-    "cf:deploy:forwarder:stg": "wrangler deploy --env stg --config workers/posthog-log-forwarder/wrangler.jsonc",
-    "knip": "knip"
+    "cf:deploy:prd": "bun run cf:typegen && CLOUDFLARE_ENV=production bun run build && wrangler deploy --env=production"
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.1047.0",
