+++ scripts/env-file.ts
@@ -1,0 +1,124 @@
+/**
+ * Shared .env file helpers for the setup/bootstrap scripts
+ * (ensure-env.ts, setup.ts, setup-prod.ts).
+ */
+
+import { execFileSync } from 'child_process';
+import { existsSync, readFileSync, writeFileSync } from 'fs';
+import { resolve } from 'path';
+
+export function parseEnvString(text: string): Map<string, string> {
+  const env = new Map<string, string>();
+  for (const line of text.split('\n')) {
+    const trimmed = line.trim();
+    if (!trimmed || trimmed.startsWith('#')) continue;
+    const eqIdx = trimmed.indexOf('=');
+    if (eqIdx === -1) continue;
+    const key = trimmed.slice(0, eqIdx).trim();
+    let value = trimmed.slice(eqIdx + 1).trim();
+    // Strip surrounding quotes (some dashboards export KEY="value")
+    if (
+      value.length >= 2 &&
+      ((value[0] === '"' && value.at(-1) === '"') ||
+        (value[0] === "'" && value.at(-1) === "'"))
+    ) {
+      value = value.slice(1, -1);
+    }
+    if (key && value) env.set(key, value);
+  }
+  return env;
+}
+
+export function parseEnvFile(path: string): Map<string, string> {
+  if (!existsSync(path)) return new Map();
+  return parseEnvString(readFileSync(path, 'utf-8'));
+}
+
+/**
+ * Set or append KEY=VALUE lines in an env file, preserving everything else
+ * (comments, ordering, keys we don't touch). Creates the file if missing.
+ */
+export function upsertEnvVars(
+  path: string,
+  vars: Record<string, string>
+): void {
+  const text = existsSync(path) ? readFileSync(path, 'utf-8') : '';
+  const lines = text.length > 0 ? text.split('\n') : [];
+  const remaining = new Map(Object.entries(vars));
+
+  const updated = lines.map((line) => {
+    const trimmed = line.trim();
+    if (!trimmed || trimmed.startsWith('#')) return line;
+    const eqIdx = trimmed.indexOf('=');
+    if (eqIdx === -1) return line;
+    const key = trimmed.slice(0, eqIdx).trim();
+    const value = remaining.get(key);
+    if (value === undefined) return line;
+    remaining.delete(key);
+    return `${key}=${value}`;
+  });
+
+  if (remaining.size > 0) {
+    if (updated.length > 0 && updated.at(-1) !== '') updated.push('');
+    for (const [key, value] of remaining) {
+      updated.push(`${key}=${value}`);
+    }
+    updated.push('');
+  }
+
+  writeFileSync(path, updated.join('\n'));
+}
+
+const LOCAL_ENV_HEADER = `# =============================================================================
+# Local Development Configuration
+# Generated by \`bun dev\` — safe to edit. See .env.example for all options.
+# Add FAL_KEY and OPENROUTER_KEY (or run \`bun setup\`) to enable AI generation.
+# =============================================================================
+`;
+
+/**
+ * Zero-touch local env bootstrap. Creates .env.local if missing and fills in
+ * the few values the app cannot default at runtime (auth/encryption secrets
+ * are generated; VITE_APP_URL is needed by the local /r2 serve route). Never
+ * overwrites existing values. Returns the keys it added.
+ */
+export function ensureLocalEnv(): string[] {
+  const envFile = resolve(process.cwd(), '.env.local');
+  const existing = parseEnvFile(envFile);
+
+  const required: Record<string, () => string> = {
+    VITE_APP_URL: () => 'http://localhost:3000',
+    BETTER_AUTH_SECRET: generateSecret,
+    API_KEY_ENCRYPTION_KEY: generateSecret,
+  };
+
+  const missing: Record<string, string> = {};
+  for (const [key, makeValue] of Object.entries(required)) {
+    if (!existing.has(key)) missing[key] = makeValue();
+  }
+
+  const added = Object.keys(missing);
+  if (added.length === 0) return added;
+
+  const isNewFile = !existsSync(envFile);
+  upsertEnvVars(envFile, missing);
+  if (isNewFile) {
+    writeFileSync(envFile, LOCAL_ENV_HEADER + readFileSync(envFile, 'utf-8'));
+  }
+
+  return added;
+}
+
+export function generateSecret(): string {
+  try {
+    return execFileSync('openssl', ['rand', '-hex', '32'], {
+      encoding: 'utf-8',
+    }).trim();
+  } catch {
+    const bytes = new Uint8Array(32);
+    crypto.getRandomValues(bytes);
+    return Array.from(bytes)
+      .map((b) => b.toString(16).padStart(2, '0'))
+      .join('');
+  }
+}
