+++ scripts/setup-prod.ts
@@ -1,0 +1,1778 @@
+/**
+ * Production & deployment setup for OpenStory — invoked via `bun setup`:
+ *
+ *   bun setup --prod        configure .env.production interactively + deploy
+ *   bun setup --deploy      push secrets + deploy using existing .env.production
+ *   bun setup --pr-preview  derive preview secrets and push to the GitHub
+ *                           "staging" environment used by PR preview deploys
+ *
+ * Local development needs none of this — see scripts/setup.ts.
+ */
+
+import * as p from '@clack/prompts';
+import chalk from 'chalk';
+import { execFileSync } from 'child_process';
+import { writeFileSync } from 'fs';
+import { resolve } from 'path';
+import { generateSecret, parseEnvFile, parseEnvString } from './env-file';
+import {
+  buildRules,
+  derivePreviewOriginPatterns,
+  setBucketCors,
+} from './r2-cors';
+
+const ENV_FILE = resolve(process.cwd(), '.env.production');
+const ENV_FILENAME = '.env.production';
+
+export type ProdSetupMode = 'prod' | 'deploy' | 'pr-preview';
+
+// ---------------------------------------------------------------------------
+// Helpers
+// ---------------------------------------------------------------------------
+
+function loadExistingEnv(): Map<string, string> {
+  const env = parseEnvFile(ENV_FILE);
+
+  // Also read legacy .env so existing keys are migrated
+  const legacy = parseEnvFile(resolve(process.cwd(), '.env'));
+  legacy.forEach((value, key) => {
+    if (!env.has(key)) env.set(key, value);
+  });
+
+  return env;
+}
+
+function writeEnvFile(vars: Map<string, string>) {
+  const lines: string[] = [
+    '# =============================================================================',
+    '# Production Configuration',
+    '# Generated by `bun setup --prod` — safe to edit',
+    '# =============================================================================',
+    '',
+  ];
+
+  const sections: Array<{ header: string; keys: string[] }> = [
+    {
+      header: 'Core',
+      keys: [
+        'VITE_APP_URL',
+        'VITE_APP_NAME',
+        'BETTER_AUTH_SECRET',
+        'CLOUDFLARE_ACCOUNT_ID',
+        'CLOUDFLARE_API_TOKEN',
+        'CLOUDFLARE_ZONE_ID',
+      ],
+    },
+    {
+      header: 'Email (Cloudflare Email Service)',
+      keys: ['EMAIL_FROM'],
+    },
+    {
+      header: 'AI Services',
+      keys: ['FAL_KEY', 'OPENROUTER_KEY'],
+    },
+    {
+      header: 'Storage (Cloudflare R2)',
+      keys: [
+        'R2_ACCOUNT_ID',
+        'R2_ACCESS_KEY_ID',
+        'R2_SECRET_ACCESS_KEY',
+        'R2_BUCKET_NAME',
+        'R2_PUBLIC_STORAGE_DOMAIN',
+        'R2_PUBLIC_ASSETS_BUCKET',
+        'VITE_R2_PUBLIC_ASSETS_DOMAIN',
+      ],
+    },
+    {
+      header: 'Google OAuth',
+      keys: ['GOOGLE_CLIENT_ID', 'GOOGLE_CLIENT_SECRET'],
+    },
+    {
+      header: 'Observability (Langfuse)',
+      keys: [
+        'LANGFUSE_PUBLIC_KEY',
+        'LANGFUSE_SECRET_KEY',
+        'LANGFUSE_BASE_URL',
+        'LANGFUSE_PROMPTS_ENABLED',
+      ],
+    },
+    {
+      header: 'Analytics (PostHog)',
+      keys: ['VITE_PUBLIC_POSTHOG_PROJECT_TOKEN', 'VITE_PUBLIC_POSTHOG_HOST'],
+    },
+    {
+      header: 'Billing (Stripe)',
+      keys: ['STRIPE_SECRET_KEY', 'STRIPE_WEBHOOK_SECRET'],
+    },
+    {
+      header: 'Security',
+      keys: ['API_KEY_ENCRYPTION_KEY'],
+    },
+  ];
+
+  for (const section of sections) {
+    const sectionVars = section.keys.filter((k) => vars.has(k));
+    if (sectionVars.length === 0) continue;
+
+    lines.push(`# ${section.header}`);
+    for (const key of sectionVars) {
+      lines.push(`${key}=${vars.get(key)}`);
+    }
+    lines.push('');
+  }
+
+  writeFileSync(ENV_FILE, lines.join('\n'));
+}
+
+function commandExists(cmd: string): boolean {
+  try {
+    execFileSync('which', [cmd], { stdio: 'ignore' });
+    return true;
+  } catch {
+    return false;
+  }
+}
+
+// ---------------------------------------------------------------------------
+// PR Preview Setup — push secrets to GitHub staging environment
+// ---------------------------------------------------------------------------
+
+const PR_PREVIEW_SECRETS_BASE = [
+  'API_KEY_ENCRYPTION_KEY',
+  'VITE_APP_NAME',
+  'BETTER_AUTH_SECRET',
+  'EMAIL_FROM',
+  'FAL_KEY',
+  'LANGFUSE_BASE_URL',
+  'LANGFUSE_PROMPTS_ENABLED',
+  'LANGFUSE_PUBLIC_KEY',
+  'LANGFUSE_SECRET_KEY',
+  'OPENROUTER_KEY',
+  'R2_ACCESS_KEY_ID',
+  'R2_ACCOUNT_ID',
+  'R2_BUCKET_NAME',
+  'R2_PUBLIC_ASSETS_BUCKET',
+  'R2_SECRET_ACCESS_KEY',
+  'VITE_PUBLIC_POSTHOG_PROJECT_TOKEN',
+] as const;
+
+function getPrPreviewSecrets(): string[] {
+  return [
+    ...PR_PREVIEW_SECRETS_BASE,
+    'CLOUDFLARE_ACCOUNT_ID',
+    'CLOUDFLARE_API_TOKEN',
+  ];
+}
+
+const PR_PREVIEW_VARIABLES = [
+  'VITE_PUBLIC_POSTHOG_HOST',
+  'VITE_R2_PUBLIC_ASSETS_DOMAIN',
+  'R2_PUBLIC_STORAGE_DOMAIN',
+] as const;
+
+function execGh(args: string[]): string {
+  return execFileSync('gh', args, {
+    encoding: 'utf-8',
+    stdio: ['pipe', 'pipe', 'pipe'],
+  }).trim();
+}
+
+function bucketHasDomain(bucket: string, domain: string): boolean {
+  try {
+    const output = execFileSync(
+      'bunx',
+      ['wrangler', 'r2', 'bucket', 'domain', 'list', bucket],
+      { stdio: ['pipe', 'pipe', 'pipe'], encoding: 'utf-8' }
+    );
+    return output.includes(domain);
+  } catch {
+    return false;
+  }
+}
+
+async function getCloudflareWorkersSubdomain(
+  accountId: string,
+  apiToken: string
+): Promise<string | undefined> {
+  try {
+    const res = await fetch(
+      `https://api.cloudflare.com/client/v4/accounts/${accountId}/workers/subdomain`,
+      { headers: { Authorization: `Bearer ${apiToken}` } }
+    );
+    const data: { success?: boolean; result?: { name?: string } } =
+      await res.json();
+    if (data.success && data.result?.name) return data.result.name;
+  } catch {
+    // fall through
+  }
+  return undefined;
+}
+
+async function prPreviewSetup() {
+  p.intro(chalk.bold('OpenStory PR Preview Setup'));
+  p.log.info(
+    'Derives preview secrets from .env.production and pushes to GitHub "staging" environment.'
+  );
+
+  // 1. Check gh CLI
+  if (!commandExists('gh')) {
+    p.log.error(
+      'GitHub CLI (gh) is not installed.\n' +
+        'Install it: https://cli.github.com/'
+    );
+    process.exit(1);
+  }
+
+  try {
+    execGh(['auth', 'status']);
+  } catch {
+    p.log.error('Not authenticated with gh. Run: gh auth login');
+    process.exit(1);
+  }
+  p.log.success('GitHub CLI authenticated');
+
+  // 2. Detect repo
+  let repoName: string;
+  try {
+    const repoJson = execGh([
+      'repo',
+      'view',
+      '--json',
+      'nameWithOwner',
+      '-q',
+      '.nameWithOwner',
+    ]);
+    repoName = repoJson;
+  } catch {
+    p.log.error(
+      'Could not detect GitHub repository. Are you in a git repo with a GitHub remote?'
+    );
+    process.exit(1);
+  }
+  p.log.success(`Repository: ${repoName}`);
+
+  // 3. Load .env.production as base, then overlay .env.staging if it exists
+  const prodFile = resolve(process.cwd(), '.env.production');
+  const prodVars = parseEnvFile(prodFile);
+
+  if (prodVars.size === 0) {
+    p.log.error(
+      'No .env.production found. Run `bun setup --prod` first to configure production values.'
+    );
+    process.exit(1);
+  }
+
+  const stagingFile = resolve(process.cwd(), '.env.staging');
+  const stagingVars = parseEnvFile(stagingFile);
+
+  const merged = new Map(prodVars);
+  // .env.staging values take precedence over .env.production
+  for (const [key, value] of stagingVars) {
+    merged.set(key, value);
+  }
+
+  if (stagingVars.size > 0) {
+    p.log.success(
+      `Loaded ${prodVars.size} values from .env.production, ${stagingVars.size} overrides from .env.staging`
+    );
+  } else {
+    p.log.info(`Loaded ${prodVars.size} values from .env.production`);
+  }
+
+  // 3b. Fallback to .env.local for account-level R2 credentials
+  const localFile = resolve(process.cwd(), '.env.local');
+  const localVars = parseEnvFile(localFile);
+  const R2_FALLBACK_KEYS = ['R2_ACCESS_KEY_ID', 'R2_SECRET_ACCESS_KEY'];
+  for (const key of R2_FALLBACK_KEYS) {
+    const localVal = localVars.get(key);
+    if (!merged.get(key) && localVal) {
+      merged.set(key, localVal);
+    }
+  }
+
+  // 3c. Auto-derive R2_ACCOUNT_ID from CLOUDFLARE_ACCOUNT_ID (same value)
+  const cfAccountId = merged.get('CLOUDFLARE_ACCOUNT_ID');
+  if (!merged.get('R2_ACCOUNT_ID') && cfAccountId) {
+    merged.set('R2_ACCOUNT_ID', cfAccountId);
+  }
+
+  // 4. Auto-set preview overrides
+  const removed = [
+    'VITE_APP_URL',
+    'GOOGLE_CLIENT_ID',
+    'GOOGLE_CLIENT_SECRET',
+    'STRIPE_SECRET_KEY',
+    'STRIPE_WEBHOOK_SECRET',
+  ];
+  for (const key of removed) {
+    merged.delete(key);
+  }
+  p.log.info(`Removed (not needed for preview): ${removed.join(', ')}`);
+
+  // 5. Prompt for AI keys — skip if .env.staging already has its own values
+  const aiKeys = ['FAL_KEY', 'OPENROUTER_KEY'] as const;
+  const stagingHasAiKeys = aiKeys.every((k) => stagingVars.get(k));
+  const hasAiKeys = aiKeys.every((k) => merged.get(k));
+
+  if (stagingHasAiKeys) {
+    p.log.success('AI keys already configured in .env.staging');
+  } else if (hasAiKeys) {
+    const reuseProd = await p.confirm({
+      message: 'Use the same FAL_KEY and OPENROUTER_KEY as production?',
+      initialValue: true,
+    });
+
+    if (p.isCancel(reuseProd)) {
+      p.cancel('PR preview setup cancelled.');
+      process.exit(0);
+    }
+
+    if (!reuseProd) {
+      for (const key of aiKeys) {
+        const value = await p.text({
+          message: `${key} for preview`,
+          placeholder: `Paste your ${key} here…`,
+        });
+
+        if (p.isCancel(value)) {
+          p.cancel('PR preview setup cancelled.');
+          process.exit(0);
+        }
+
+        if (value.trim()) {
+          merged.set(key, value.trim());
+        }
+      }
+    }
+  }
+
+  // 6. Derive staging R2 domains from production VITE_APP_URL
+  const appUrl = prodVars.get('VITE_APP_URL') ?? '';
+  let baseDomain = '';
+  try {
+    const host = new URL(appUrl).hostname;
+    const parts = host.split('.');
+    baseDomain = parts.length >= 2 ? parts.slice(-2).join('.') : host;
+  } catch {
+    // fall through
+  }
+
+  if (baseDomain) {
+    merged.set('R2_PUBLIC_STORAGE_DOMAIN', `storage-stg.${baseDomain}`);
+    merged.set('VITE_R2_PUBLIC_ASSETS_DOMAIN', `assets.${baseDomain}`);
+  } else {
+    p.log.warn(
+      'Could not derive R2 domains from VITE_APP_URL. You may need to set R2_PUBLIC_STORAGE_DOMAIN and VITE_R2_PUBLIC_ASSETS_DOMAIN manually in GitHub.'
+    );
+  }
+
+  // 6b. Derive staging bucket names (with -stg suffix, from production values only)
+  const prodStorageBucket =
+    prodVars.get('R2_BUCKET_NAME') ?? 'openstory-storage';
+  const prodAssetsBucket =
+    prodVars.get('R2_PUBLIC_ASSETS_BUCKET') ?? 'openstory-public-assets';
+  merged.set('R2_BUCKET_NAME', `${prodStorageBucket}-stg`);
+  merged.set('R2_PUBLIC_ASSETS_BUCKET', prodAssetsBucket);
+
+  p.log.info(
+    `R2_PUBLIC_STORAGE_DOMAIN = ${merged.get('R2_PUBLIC_STORAGE_DOMAIN') ?? '(not set)'}\n` +
+      `VITE_R2_PUBLIC_ASSETS_DOMAIN  = ${merged.get('VITE_R2_PUBLIC_ASSETS_DOMAIN') ?? '(not set)'}`
+  );
+
+  // Offer to attach custom domains to R2 buckets
+  const stgStorageDomain = merged.get('R2_PUBLIC_STORAGE_DOMAIN');
+  const stgAssetsDomain = merged.get('VITE_R2_PUBLIC_ASSETS_DOMAIN');
+
+  if (stgStorageDomain || stgAssetsDomain) {
+    const attachDomains = await p.confirm({
+      message: 'Attach these custom domains to R2 buckets via Cloudflare?',
+      initialValue: true,
+    });
+
+    if (p.isCancel(attachDomains)) {
+      p.cancel('PR preview setup cancelled.');
+      process.exit(0);
+    }
+
+    if (attachDomains) {
+      // Auto-detect zone ID from domain
+      let zoneId = prodVars.get('CLOUDFLARE_ZONE_ID');
+      const apiToken = prodVars.get('CLOUDFLARE_API_TOKEN');
+
+      if (!zoneId && apiToken && baseDomain) {
+        const zoneSpinner = p.spinner();
+        zoneSpinner.start(`Looking up zone ID for ${baseDomain}`);
+        try {
+          const res = await fetch(
+            `https://api.cloudflare.com/client/v4/zones?name=${baseDomain}`,
+            { headers: { Authorization: `Bearer ${apiToken}` } }
+          );
+          const data: { success?: boolean; result?: { id: string }[] } =
+            await res.json();
+          const firstZone = data.success ? data.result?.[0] : undefined;
+          if (firstZone) {
+            zoneId = firstZone.id;
+            zoneSpinner.stop(
+              `Found zone ID for ${baseDomain}: ${zoneId.slice(0, 8)}…`
+            );
+          } else {
+            zoneSpinner.stop(`Could not find zone for ${baseDomain}`);
+          }
+        } catch {
+          zoneSpinner.stop('Zone lookup failed');
+        }
+      }
+
+      if (!zoneId) {
+        const manualZoneId = await p.text({
+          message:
+            'Cloudflare Zone ID\n' +
+            chalk.dim('Find in Cloudflare Dashboard > your domain > Overview'),
+          placeholder: 'Paste your CLOUDFLARE_ZONE_ID here…',
+        });
+
+        if (p.isCancel(manualZoneId)) {
+          p.cancel('PR preview setup cancelled.');
+          process.exit(0);
+        }
+
+        zoneId = manualZoneId.trim() || undefined;
+      }
+
+      if (zoneId) {
+        // Use already-derived staging bucket names (set in step 6b)
+        const storageBucket =
+          merged.get('R2_BUCKET_NAME') ?? 'openstory-storage-stg';
+        const assetsBucket =
+          merged.get('R2_PUBLIC_ASSETS_BUCKET') ?? 'openstory-public-assets';
+
+        for (const [domain, bucket] of [
+          [stgStorageDomain, storageBucket],
+          [stgAssetsDomain, assetsBucket],
+        ] as const) {
+          if (!domain) continue;
+          if (bucketHasDomain(bucket, domain)) {
+            p.log.success(`${domain} → ${bucket} (already attached)`);
+            continue;
+          }
+          try {
+            execFileSync(
+              'bunx',
+              [
+                'wrangler',
+                'r2',
+                'bucket',
+                'domain',
+                'add',
+                bucket,
+                '--domain',
+                domain,
+                '--zone-id',
+                zoneId,
+                '--force',
+              ],
+              { stdio: 'inherit' }
+            );
+            p.log.success(`Attached ${domain} → ${bucket}`);
+          } catch {
+            p.log.warn(`Could not attach ${domain} to ${bucket}`);
+          }
+        }
+      } else {
+        p.log.warn('No zone ID — skipping domain attachment.');
+      }
+    }
+  }
+
+  // 6c. Configure CORS on the staging bucket. Each PR deploy gets a unique
+  // workers.dev URL so we use a wildcard origin matching the preview pattern.
+  // GET/HEAD only — staging buckets don't accept browser uploads.
+  const stgBucketName = merged.get('R2_BUCKET_NAME');
+  if (stgBucketName) {
+    let workersSubdomain: string | undefined;
+    const accountId = merged.get('CLOUDFLARE_ACCOUNT_ID');
+    const apiToken = merged.get('CLOUDFLARE_API_TOKEN');
+    if (accountId && apiToken) {
+      const subSpinner = p.spinner();
+      subSpinner.start('Looking up Cloudflare Workers subdomain');
+      workersSubdomain = await getCloudflareWorkersSubdomain(
+        accountId,
+        apiToken
+      );
+      if (workersSubdomain) {
+        subSpinner.stop(`Workers subdomain: ${workersSubdomain}`);
+      } else {
+        subSpinner.stop('Could not auto-detect Workers subdomain');
+      }
+    }
+
+    if (!workersSubdomain) {
+      const manual = await p.text({
+        message:
+          'Workers subdomain (the part before .workers.dev in your preview URL)',
+        placeholder: 'e.g. myaccount',
+      });
+      if (!p.isCancel(manual) && manual.trim()) {
+        workersSubdomain = manual.trim();
+      }
+    }
+
+    const previewOrigins = derivePreviewOriginPatterns({ workersSubdomain });
+
+    if (previewOrigins.length === 0) {
+      p.log.warn(
+        `No Workers subdomain available.\n` +
+          `Configure CORS manually in Cloudflare Dashboard:\n` +
+          `  R2 → ${stgBucketName} → Settings → CORS Policy`
+      );
+    } else {
+      const setupStgCors = await p.confirm({
+        message: `Configure CORS on staging bucket ${chalk.bold(stgBucketName)}?`,
+        initialValue: true,
+      });
+
+      if (p.isCancel(setupStgCors)) {
+        p.cancel('PR preview setup cancelled.');
+        process.exit(0);
+      }
+
+      if (setupStgCors) {
+        const rules = buildRules({
+          origins: previewOrigins,
+          includeWrites: false,
+        });
+        const corsSpinner = p.spinner();
+        corsSpinner.start(`Configuring CORS on ${stgBucketName}`);
+        try {
+          setBucketCors(stgBucketName, rules);
+          corsSpinner.stop(
+            `CORS configured on ${stgBucketName} (${previewOrigins.join(', ')})`
+          );
+        } catch {
+          corsSpinner.stop('CORS configuration failed');
+          p.log.warn(
+            'Configure manually in Cloudflare Dashboard:\n' +
+              `  R2 → ${stgBucketName} → Settings → CORS Policy\n\n` +
+              JSON.stringify({ rules }, null, 2)
+          );
+        }
+      }
+    }
+  }
+
+  // 7. Summary + confirm
+  const previewSecrets = getPrPreviewSecrets();
+  const allKeys = [...previewSecrets, ...PR_PREVIEW_VARIABLES];
+  const ready = allKeys.filter((k) => merged.get(k));
+  const missing = allKeys.filter((k) => !merged.get(k));
+
+  p.log.info(
+    `${chalk.green(String(ready.length))} keys ready, ${chalk.yellow(String(missing.length))} missing`
+  );
+
+  if (missing.length > 0) {
+    p.log.warn(`Missing: ${missing.join(', ')}`);
+  }
+
+  // Confirm before pushing to GitHub
+  const confirmPush = await p.confirm({
+    message: `Push ${ready.length} keys to GitHub "staging" environment?`,
+    initialValue: true,
+  });
+
+  if (p.isCancel(confirmPush) || !confirmPush) {
+    p.cancel('PR preview setup cancelled.');
+    process.exit(0);
+  }
+
+  // 8. Save .env.staging locally
+  const stagingLines = [
+    '# =============================================================================',
+    '# Staging Configuration',
+    '# Generated by `bun setup --pr-preview` — safe to edit',
+    '# =============================================================================',
+    '',
+  ];
+  for (const [key, value] of merged) {
+    stagingLines.push(`${key}=${value}`);
+  }
+  stagingLines.push('');
+  writeFileSync(stagingFile, stagingLines.join('\n'));
+  p.log.success(`Wrote ${stagingFile}`);
+
+  // 9. Push to GitHub staging environment
+  const envSpinner = p.spinner();
+  envSpinner.start('Ensuring GitHub "staging" environment exists');
+  try {
+    execGh([
+      'api',
+      '--method',
+      'PUT',
+      `/repos/${repoName}/environments/staging`,
+    ]);
+    envSpinner.stop('GitHub "staging" environment ready');
+  } catch (error) {
+    envSpinner.stop('Failed to create staging environment');
+    p.log.error(
+      `Error: ${error instanceof Error ? error.message : String(error)}\n` +
+        'You may need to create it manually in GitHub Settings > Environments.'
+    );
+    process.exit(1);
+  }
+
+  // Push secrets
+  const secretSpinner = p.spinner();
+  const secretsToPush = previewSecrets.filter((k) => merged.get(k));
+  secretSpinner.start(
+    `Pushing ${secretsToPush.length} secrets to staging environment`
+  );
+  let secretsPushed = 0;
+  let secretsFailed = 0;
+
+  for (const key of secretsToPush) {
+    const value = merged.get(key);
+    if (!value) continue;
+    try {
+      execFileSync(
+        'gh',
+        ['secret', 'set', key, '--env', 'staging', '--body', value],
+        {
+          stdio: ['pipe', 'pipe', 'pipe'],
+        }
+      );
+      secretsPushed++;
+    } catch {
+      secretsFailed++;
+      p.log.warn(`Failed to push secret: ${key}`);
+    }
+  }
+
+  if (secretsFailed > 0) {
+    secretSpinner.stop(
+      `Pushed ${secretsPushed} secrets (${secretsFailed} failed)`
+    );
+  } else {
+    secretSpinner.stop(`Pushed ${secretsPushed} secrets`);
+  }
+
+  // Push variables
+  const varsToPush = PR_PREVIEW_VARIABLES.filter((k) => merged.get(k));
+  if (varsToPush.length > 0) {
+    const varSpinner = p.spinner();
+    varSpinner.start(
+      `Pushing ${varsToPush.length} variables to staging environment`
+    );
+    let varsPushed = 0;
+    let varsFailed = 0;
+
+    for (const key of varsToPush) {
+      const value = merged.get(key);
+      if (!value) continue;
+      try {
+        execFileSync(
+          'gh',
+          ['variable', 'set', key, '--env', 'staging', '--body', value],
+          {
+            stdio: ['pipe', 'pipe', 'pipe'],
+          }
+        );
+        varsPushed++;
+      } catch {
+        varsFailed++;
+        p.log.warn(`Failed to push variable: ${key}`);
+      }
+    }
+
+    if (varsFailed > 0) {
+      varSpinner.stop(`Pushed ${varsPushed} variables (${varsFailed} failed)`);
+    } else {
+      varSpinner.stop(`Pushed ${varsPushed} variables`);
+    }
+  }
+
+  // 9. Summary
+  p.note(
+    [
+      `Repository: ${repoName}`,
+      `Environment: staging`,
+      `Secrets pushed: ${secretsPushed}`,
+      `Variables pushed: ${varsToPush.length}`,
+    ].join('\n'),
+    'PR Preview Setup Complete'
+  );
+
+  p.outro('Done! PR preview deploys will now use the staging environment.');
+}
+
+// ---------------------------------------------------------------------------
+// Deploy Setup
+// ---------------------------------------------------------------------------
+
+const config = {
+  cmd: 'wrangler',
+  install: 'bun add -g wrangler',
+  loginArgs: ['login'] as string[],
+  whoamiArgs: ['whoami'] as string[],
+};
+
+async function deploySetup(
+  vars: Map<string, string>,
+  checkCancel: <T>(value: T | symbol) => T
+) {
+  p.log.step(chalk.bold('Deploy Setup — Cloudflare Workers'));
+
+  // 1. Check CLI installed
+  if (!commandExists(config.cmd)) {
+    const shouldInstall = checkCancel(
+      await p.confirm({
+        message: `${config.cmd} is not installed. Install it now?`,
+        initialValue: true,
+      })
+    );
+
+    if (shouldInstall) {
+      const installSpinner = p.spinner();
+      installSpinner.start(`Installing ${config.cmd}`);
+      try {
+        const [cmd, ...args] = config.install.split(' ');
+        if (!cmd) {
+          throw new Error(`Invalid install command: ${config.install}`);
+        }
+        execFileSync(cmd, args, { stdio: 'pipe' });
+        installSpinner.stop(`${config.cmd} installed`);
+      } catch (error) {
+        installSpinner.stop(`Failed to install ${config.cmd}`);
+        p.log.error(
+          `Install error: ${error instanceof Error ? error.message : String(error)}`
+        );
+        p.log.info(`Run manually: ${chalk.bold(config.install)}`);
+        return;
+      }
+    } else {
+      p.log.info(`Install manually: ${chalk.bold(config.install)}`);
+      return;
+    }
+  } else {
+    p.log.success(`${config.cmd} installed`);
+  }
+
+  // 2. Check auth
+  let isAuthed = false;
+  try {
+    execFileSync(config.cmd, config.whoamiArgs, { stdio: 'pipe' });
+    isAuthed = true;
+  } catch {
+    isAuthed = false;
+  }
+
+  if (!isAuthed) {
+    const shouldLogin = checkCancel(
+      await p.confirm({
+        message: `Not logged in to ${config.cmd}. Log in now?`,
+        initialValue: true,
+      })
+    );
+
+    if (shouldLogin) {
+      p.log.info('Opening browser for authentication…');
+      try {
+        execFileSync(config.cmd, config.loginArgs, { stdio: 'inherit' });
+      } catch (error) {
+        p.log.error(
+          `Login failed: ${error instanceof Error ? error.message : String(error)}`
+        );
+        return;
+      }
+
+      // Re-check
+      try {
+        execFileSync(config.cmd, config.whoamiArgs, { stdio: 'pipe' });
+        p.log.success(`Logged in to ${config.cmd}`);
+      } catch {
+        p.log.error(
+          `Still not authenticated. Run ${chalk.bold(`${config.cmd} login`)} manually.`
+        );
+        return;
+      }
+    } else {
+      p.log.info(`Run ${chalk.bold(`${config.cmd} login`)} when ready.`);
+      return;
+    }
+  } else {
+    p.log.success(`Authenticated with ${config.cmd}`);
+  }
+
+  // 3. Push secrets to Cloudflare Workers
+  {
+    const shouldPushSecrets = checkCancel(
+      await p.confirm({
+        message: 'Push secrets to Cloudflare Workers?',
+        initialValue: true,
+      })
+    );
+
+    if (shouldPushSecrets) {
+      const secretsSpinner = p.spinner();
+      secretsSpinner.start('Pushing secrets to Cloudflare');
+      try {
+        execFileSync('wrangler', ['secret', 'bulk', ENV_FILE], {
+          stdio: 'pipe',
+        });
+        secretsSpinner.stop('Secrets pushed to Cloudflare');
+      } catch (error) {
+        secretsSpinner.stop('Failed to push secrets');
+        p.log.error(
+          `Error: ${error instanceof Error ? error.message : String(error)}`
+        );
+        p.log.info(
+          `Run manually: ${chalk.bold(`wrangler secret bulk ${ENV_FILENAME}`)}`
+        );
+      }
+    }
+
+    // Push R2 domain variables to GitHub production environment
+    // (CI seed step reads these via vars.VITE_R2_PUBLIC_ASSETS_DOMAIN)
+    const varsToPush = PR_PREVIEW_VARIABLES.filter((k) => vars.get(k));
+    if (varsToPush.length > 0) {
+      const shouldPushVars = checkCancel(
+        await p.confirm({
+          message:
+            'Push variables to GitHub production environment? (needed for CI build + seed)',
+          initialValue: true,
+        })
+      );
+
+      if (shouldPushVars) {
+        const varSpinner = p.spinner();
+        varSpinner.start(
+          `Pushing ${varsToPush.length} variables to production environment`
+        );
+
+        try {
+          const repoName = execGh([
+            'repo',
+            'view',
+            '--json',
+            'nameWithOwner',
+            '-q',
+            '.nameWithOwner',
+          ]);
+
+          // Ensure the production environment exists
+          execGh([
+            'api',
+            '--method',
+            'PUT',
+            `/repos/${repoName}/environments/production`,
+          ]);
+
+          let varsPushed = 0;
+          let varsFailed = 0;
+
+          for (const key of varsToPush) {
+            const value = vars.get(key);
+            if (!value) continue;
+            try {
+              execGh([
+                'variable',
+                'set',
+                key,
+                '--env',
+                'production',
+                '--body',
+                value,
+              ]);
+              varsPushed++;
+            } catch {
+              varsFailed++;
+              p.log.warn(`Failed to push variable: ${key}`);
+            }
+          }
+
+          if (varsFailed > 0) {
+            varSpinner.stop(
+              `Pushed ${varsPushed} variables (${varsFailed} failed)`
+            );
+          } else {
+            varSpinner.stop(`Pushed ${varsPushed} variables`);
+          }
+        } catch (error) {
+          varSpinner.stop('Failed to push variables');
+          p.log.error(
+            `Error: ${error instanceof Error ? error.message : String(error)}`
+          );
+          p.log.info(
+            'Run manually:\n' +
+              varsToPush
+                .map(
+                  (k) =>
+                    `  gh variable set ${k} --env production --body "${vars.get(k)}"`
+                )
+                .join('\n')
+          );
+        }
+      }
+    }
+  }
+
+  // 4. Push CI secrets to GitHub production environment
+  const ciSecrets = [
+    'CLOUDFLARE_ACCOUNT_ID',
+    'CLOUDFLARE_API_TOKEN',
+    'VITE_PUBLIC_POSTHOG_PROJECT_TOKEN',
+  ] as const;
+  const ciSecretsToPush = ciSecrets.filter((k) => vars.get(k));
+
+  if (ciSecretsToPush.length > 0) {
+    const shouldPushCiSecrets = checkCancel(
+      await p.confirm({
+        message: `Push CI/build secrets (${ciSecretsToPush.join(', ')}) to GitHub production environment?`,
+        initialValue: true,
+      })
+    );
+
+    if (shouldPushCiSecrets) {
+      const ciSpinner = p.spinner();
+      ciSpinner.start(
+        `Pushing ${ciSecretsToPush.length} CI secrets to production environment`
+      );
+
+      try {
+        const repoName = execGh([
+          'repo',
+          'view',
+          '--json',
+          'nameWithOwner',
+          '-q',
+          '.nameWithOwner',
+        ]);
+
+        // Ensure the production environment exists
+        execGh([
+          'api',
+          '--method',
+          'PUT',
+          `/repos/${repoName}/environments/production`,
+        ]);
+
+        let pushed = 0;
+        let failed = 0;
+
+        for (const key of ciSecretsToPush) {
+          const value = vars.get(key);
+          if (!value) continue;
+          try {
+            execGh([
+              'secret',
+              'set',
+              key,
+              '--env',
+              'production',
+              '--body',
+              value,
+            ]);
+            pushed++;
+          } catch {
+            failed++;
+            p.log.warn(`Failed to push CI secret: ${key}`);
+          }
+        }
+
+        if (failed > 0) {
+          ciSpinner.stop(`Pushed ${pushed} CI secrets (${failed} failed)`);
+        } else {
+          ciSpinner.stop(`Pushed ${pushed} CI secrets`);
+        }
+      } catch (error) {
+        ciSpinner.stop('Failed to push CI secrets');
+        p.log.error(
+          `Error: ${error instanceof Error ? error.message : String(error)}`
+        );
+        p.log.info(
+          'Run manually:\n' +
+            ciSecretsToPush
+              .map(
+                (k) =>
+                  `  gh secret set ${k} --env production --body "${vars.get(k)}"`
+              )
+              .join('\n')
+        );
+      }
+    }
+  }
+
+  // 5. Deploy
+  const shouldDeploy = checkCancel(
+    await p.confirm({
+      message: 'Deploy to Cloudflare Workers now?',
+      initialValue: true,
+    })
+  );
+
+  if (!shouldDeploy) {
+    p.log.info('Skipping deploy. You can deploy later manually.');
+    return;
+  }
+
+  p.log.info(
+    'D1 migrations and seeding are handled by the GitHub Actions deploy workflow.'
+  );
+
+  p.log.info('Building and deploying to Cloudflare Workers…');
+  try {
+    execFileSync('bun', ['cf:deploy:prd'], {
+      stdio: 'inherit',
+      cwd: process.cwd(),
+    });
+    p.log.success('Deployed to Cloudflare Workers');
+  } catch (error) {
+    p.log.error(
+      `Deploy failed: ${error instanceof Error ? error.message : String(error)}`
+    );
+    p.log.info(`Run manually: ${chalk.bold('bun cf:deploy:prd')}`);
+  }
+}
+
+// ---------------------------------------------------------------------------
+// Production config flow
+// ---------------------------------------------------------------------------
+
+export async function runProdSetup(mode: ProdSetupMode) {
+  const existing = loadExistingEnv();
+  const vars = new Map(existing);
+
+  // Save progress incrementally so users can cancel mid-setup
+  const saveProgress = () => writeEnvFile(vars);
+
+  process.on('SIGINT', () => {
+    saveProgress();
+    console.log('');
+    p.cancel(`Setup cancelled. Progress saved to ${ENV_FILENAME}`);
+    process.exit(0);
+  });
+
+  function checkCancel<T>(value: T | symbol): T {
+    if (p.isCancel(value)) {
+      saveProgress();
+      p.cancel(`Setup cancelled. Progress saved to ${ENV_FILENAME}`);
+      process.exit(0);
+    }
+    return value;
+  }
+
+  // --pr-preview: push secrets to GitHub staging environment for PR previews
+  if (mode === 'pr-preview') {
+    await prPreviewSetup();
+    return;
+  }
+
+  // --deploy: skip env var prompts, jump straight to deploy phase
+  if (mode === 'deploy') {
+    p.intro(chalk.bold('OpenStory Deploy Setup'));
+
+    if (existing.size === 0) {
+      p.log.error(
+        `No ${ENV_FILENAME} found. Run ${chalk.bold('bun setup --prod')} first to configure environment variables.`
+      );
+      process.exit(1);
+    }
+
+    p.log.info(`Loaded ${existing.size} values from ${ENV_FILENAME}`);
+
+    await deploySetup(vars, checkCancel);
+    p.outro('Deploy setup complete.');
+    return;
+  }
+
+  p.intro(chalk.bold('OpenStory Production Setup'));
+
+  async function promptForKey(
+    key: string,
+    message: string,
+    hint?: string
+  ): Promise<void> {
+    if (vars.has(key)) {
+      p.log.success(`${key} — already configured`);
+      return;
+    }
+
+    const raw = checkCancel(
+      await p.text({
+        message: hint ? `${message}\n${chalk.dim(hint)}` : message,
+        placeholder: `Paste your ${key} here…`,
+      })
+    );
+
+    if (!raw || !raw.trim()) return;
+
+    // Try parsing as KEY=VALUE (stores under the parsed key, not the expected key,
+    // so pasting KEY=VALUE into any prompt works correctly)
+    const parsed = parseEnvString(raw);
+    const envPairs = new Map<string, string>();
+    for (const [k, v] of parsed) {
+      if (/^[A-Z_][A-Z0-9_]*$/.test(k)) {
+        envPairs.set(k, v);
+      }
+    }
+
+    if (envPairs.size > 0) {
+      for (const [k, v] of envPairs) {
+        vars.set(k, v);
+      }
+      saveProgress();
+      return;
+    }
+
+    // Raw value without KEY= prefix — store under expected key, strip quotes
+    let value = raw.trim();
+    if (
+      value.length >= 2 &&
+      ((value[0] === '"' && value.at(-1) === '"') ||
+        (value[0] === "'" && value.at(-1) === "'"))
+    ) {
+      value = value.slice(1, -1);
+    }
+
+    vars.set(key, value);
+    saveProgress();
+  }
+
+  async function attachR2Domains(): Promise<void> {
+    const storageDomain = vars.get('R2_PUBLIC_STORAGE_DOMAIN');
+    const assetsDomain = vars.get('VITE_R2_PUBLIC_ASSETS_DOMAIN');
+
+    if (!storageDomain && !assetsDomain) return;
+
+    const attachDomains = checkCancel(
+      await p.confirm({
+        message: 'Attach custom domains to R2 buckets via Cloudflare?',
+        initialValue: true,
+      })
+    );
+
+    if (attachDomains) {
+      if (!vars.has('CLOUDFLARE_ZONE_ID')) {
+        // Try to auto-detect zone ID from the domain name
+        const domain = storageDomain ?? assetsDomain;
+        const baseDomain = domain?.split('.').slice(-2).join('.');
+        const apiToken = vars.get('CLOUDFLARE_API_TOKEN');
+
+        if (apiToken && baseDomain) {
+          const zoneSpinner = p.spinner();
+          zoneSpinner.start(`Looking up zone ID for ${baseDomain}`);
+          try {
+            const res = await fetch(
+              `https://api.cloudflare.com/client/v4/zones?name=${baseDomain}`,
+              { headers: { Authorization: `Bearer ${apiToken}` } }
+            );
+            const data: { success?: boolean; result?: { id: string }[] } =
+              await res.json();
+            const firstZone = data.success ? data.result?.[0] : undefined;
+            if (firstZone) {
+              const zoneId = firstZone.id;
+              vars.set('CLOUDFLARE_ZONE_ID', zoneId);
+              saveProgress();
+              zoneSpinner.stop(
+                `Found zone ID for ${baseDomain}: ${zoneId.slice(0, 8)}…`
+              );
+            } else {
+              zoneSpinner.stop(`Could not find zone for ${baseDomain}`);
+            }
+          } catch {
+            zoneSpinner.stop('Zone lookup failed');
+          }
+        }
+
+        if (!vars.has('CLOUDFLARE_ZONE_ID')) {
+          await promptForKey(
+            'CLOUDFLARE_ZONE_ID',
+            'Cloudflare Zone ID',
+            'Find in Cloudflare Dashboard > your domain > Overview'
+          );
+        }
+      }
+      const zoneId = vars.get('CLOUDFLARE_ZONE_ID');
+      if (!zoneId) {
+        p.log.warn('No zone ID provided — skipping domain attachment.');
+        return;
+      }
+
+      const storageBucket = vars.get('R2_BUCKET_NAME') ?? 'openstory-storage';
+      const assetsBucket =
+        vars.get('R2_PUBLIC_ASSETS_BUCKET') ?? 'openstory-public-assets';
+
+      for (const [domain, bucket] of [
+        [storageDomain, storageBucket],
+        [assetsDomain, assetsBucket],
+      ] as const) {
+        if (!domain) continue;
+        if (bucketHasDomain(bucket, domain)) {
+          p.log.success(`${domain} → ${bucket} (already attached)`);
+          continue;
+        }
+        try {
+          execFileSync(
+            'bunx',
+            [
+              'wrangler',
+              'r2',
+              'bucket',
+              'domain',
+              'add',
+              bucket,
+              '--domain',
+              domain,
+              '--zone-id',
+              zoneId,
+              '--force',
+            ],
+            { stdio: 'inherit' }
+          );
+          p.log.success(`Attached ${domain} → ${bucket}`);
+        } catch {
+          p.log.warn(`Could not attach ${domain} to ${bucket}`);
+        }
+      }
+    }
+  }
+
+  if (existing.size > 0) {
+    p.log.info(
+      `Found existing ${ENV_FILENAME} with ${existing.size} configured values`
+    );
+  }
+
+  // -------------------------------------------------------------------------
+  // Email (Cloudflare Email Service)
+  // -------------------------------------------------------------------------
+  p.log.step(chalk.bold('Email Setup'));
+  p.log.info(
+    'Email is sent via Cloudflare Email Service (the SEND_EMAIL binding).\n' +
+      'The sender domain must be onboarded under Email > Email Sending in\n' +
+      'the Cloudflare dashboard before production emails will deliver.'
+  );
+  if (!vars.has('EMAIL_FROM')) {
+    const emailFrom = checkCancel(
+      await p.text({
+        message: 'Enter sender email address',
+        placeholder: 'noreply@example.com',
+        validate: (v) =>
+          !v ? 'Sender email is required for production' : undefined,
+      })
+    );
+    if (emailFrom) {
+      vars.set('EMAIL_FROM', emailFrom);
+      saveProgress();
+    }
+  } else {
+    p.log.success('EMAIL_FROM — already configured');
+  }
+
+  // -------------------------------------------------------------------------
+  // Core
+  // -------------------------------------------------------------------------
+  p.log.step(chalk.bold('Core Setup'));
+
+  await promptForKey(
+    'VITE_APP_URL',
+    'Enter your production URL',
+    'e.g. https://app.example.com'
+  );
+
+  p.log.step(chalk.bold('Cloudflare'));
+  await promptForKey(
+    'CLOUDFLARE_ACCOUNT_ID',
+    'Enter your Cloudflare Account ID',
+    'Find in Cloudflare console → Compute → Workers & Pages → Account ID (sidebar)'
+  );
+  await promptForKey(
+    'CLOUDFLARE_API_TOKEN',
+    'Enter your Cloudflare API token',
+    'Create at: https://dash.cloudflare.com/profile/api-tokens (needs D1 + Workers edit permissions)'
+  );
+  p.log.info(
+    'D1 database is configured in wrangler.jsonc — deploy handles migrations automatically.'
+  );
+
+  if (!vars.has('VITE_APP_NAME')) vars.set('VITE_APP_NAME', 'OpenStory');
+
+  // Contact/privacy emails are derived from VITE_APP_URL at runtime
+  // (src/lib/marketing/constants.ts) — no env vars needed.
+
+  if (!vars.has('BETTER_AUTH_SECRET')) {
+    vars.set('BETTER_AUTH_SECRET', generateSecret());
+    p.log.success('Auth secret generated');
+  } else {
+    p.log.success('Auth secret — already configured');
+  }
+
+  if (!vars.has('API_KEY_ENCRYPTION_KEY')) {
+    vars.set('API_KEY_ENCRYPTION_KEY', generateSecret());
+    p.log.success('Encryption key generated');
+  } else {
+    p.log.success('Encryption key — already configured');
+  }
+
+  p.log.success('Database: Cloudflare D1 (via wrangler.jsonc)');
+  p.log.success(`App URL: ${vars.get('VITE_APP_URL')}`);
+  saveProgress();
+
+  // -------------------------------------------------------------------------
+  // AI Generation (FAL_KEY)
+  // -------------------------------------------------------------------------
+  if (vars.has('FAL_KEY')) {
+    p.log.success('FAL_KEY — already configured');
+  } else {
+    const setupAI = checkCancel(
+      await p.confirm({
+        message: 'Set up AI image & video generation?',
+        initialValue: true,
+      })
+    );
+
+    if (setupAI) {
+      await promptForKey(
+        'FAL_KEY',
+        'Enter your Fal.ai API key',
+        'Get one at: https://fal.ai/dashboard/keys'
+      );
+    }
+  }
+
+  // -------------------------------------------------------------------------
+  // Script Analysis (OPENROUTER_KEY)
+  // -------------------------------------------------------------------------
+  if (vars.has('OPENROUTER_KEY')) {
+    p.log.success('OPENROUTER_KEY — already configured');
+  } else {
+    const setupLLM = checkCancel(
+      await p.confirm({
+        message: 'Set up LLM-powered script analysis?',
+        initialValue: true,
+      })
+    );
+
+    if (setupLLM) {
+      await promptForKey(
+        'OPENROUTER_KEY',
+        'Enter your OpenRouter API key',
+        'Get one at: https://openrouter.ai/settings/keys'
+      );
+    }
+  }
+
+  // Async workflows run on Cloudflare Workflows (bindings in wrangler.jsonc).
+  // Realtime progress updates run on the `REALTIME` Durable Object. Neither
+  // needs any external service configuration.
+
+  // -------------------------------------------------------------------------
+  // Storage (R2)
+  // -------------------------------------------------------------------------
+  p.log.info(
+    'R2 storage is handled by native bindings configured in wrangler.jsonc.\n' +
+      'S3 credentials (for signed URLs) can be added to .env and deployed\n' +
+      'via `bun setup-cloudflare-secrets.sh`.'
+  );
+
+  if (
+    !vars.has('R2_PUBLIC_STORAGE_DOMAIN') ||
+    !vars.has('VITE_R2_PUBLIC_ASSETS_DOMAIN')
+  ) {
+    // Derive R2 domains from VITE_APP_URL base domain
+    const appUrl = vars.get('VITE_APP_URL') ?? '';
+    let baseDomain = '';
+    try {
+      const host = new URL(appUrl).hostname; // e.g. openstory.so
+      const parts = host.split('.');
+      baseDomain = parts.length >= 2 ? parts.slice(-2).join('.') : host;
+    } catch {
+      // VITE_APP_URL not set or invalid — fall through to manual prompt
+    }
+
+    const isStaging = checkCancel(
+      await p.select({
+        message: 'Which R2 environment is this?',
+        options: [
+          { value: 'production', label: 'Production' },
+          {
+            value: 'staging',
+            label: 'Staging / Preview',
+            hint: 'uses -stg/-dev subdomains',
+          },
+        ],
+      })
+    );
+
+    const storageSub = isStaging === 'staging' ? 'storage-stg' : 'storage';
+    const assetsSub = 'assets';
+
+    if (baseDomain) {
+      const storageDomain = `${storageSub}.${baseDomain}`;
+      const assetsDomain = `${assetsSub}.${baseDomain}`;
+
+      p.log.info(
+        `Derived from ${chalk.bold(baseDomain)}:\n` +
+          `  R2_PUBLIC_STORAGE_DOMAIN = ${storageDomain}\n` +
+          `  VITE_R2_PUBLIC_ASSETS_DOMAIN  = ${assetsDomain}`
+      );
+
+      const useDefaults = checkCancel(
+        await p.confirm({
+          message: 'Use these derived domains?',
+          initialValue: true,
+        })
+      );
+
+      if (useDefaults) {
+        vars.set('R2_PUBLIC_STORAGE_DOMAIN', storageDomain);
+        vars.set('VITE_R2_PUBLIC_ASSETS_DOMAIN', assetsDomain);
+        saveProgress();
+      } else {
+        await promptForKey(
+          'R2_PUBLIC_STORAGE_DOMAIN',
+          `R2 Public Storage Domain (e.g. ${storageSub}.example.com)`
+        );
+        await promptForKey(
+          'VITE_R2_PUBLIC_ASSETS_DOMAIN',
+          `R2 Public Assets Domain (e.g. ${assetsSub}.example.com)`
+        );
+      }
+    } else {
+      await promptForKey(
+        'R2_PUBLIC_STORAGE_DOMAIN',
+        `R2 Public Storage Domain (e.g. ${storageSub}.example.com)`
+      );
+      await promptForKey(
+        'VITE_R2_PUBLIC_ASSETS_DOMAIN',
+        `R2 Public Assets Domain (e.g. ${assetsSub}.example.com)`
+      );
+    }
+  } else {
+    p.log.success(
+      `R2_PUBLIC_STORAGE_DOMAIN = ${vars.get('R2_PUBLIC_STORAGE_DOMAIN')}`
+    );
+    p.log.success(
+      `VITE_R2_PUBLIC_ASSETS_DOMAIN  = ${vars.get('VITE_R2_PUBLIC_ASSETS_DOMAIN')}`
+    );
+  }
+
+  // Bucket names — needed for CORS configuration. Cloudflare runtime uses
+  // native bindings from wrangler.jsonc, but wrangler CLI commands like
+  // `r2 bucket cors set` need the bucket name explicitly.
+  if (!vars.has('R2_BUCKET_NAME')) {
+    vars.set('R2_BUCKET_NAME', 'openstory-storage');
+    saveProgress();
+  }
+  if (!vars.has('R2_PUBLIC_ASSETS_BUCKET')) {
+    vars.set('R2_PUBLIC_ASSETS_BUCKET', 'openstory-public-assets');
+    saveProgress();
+  }
+
+  // Configure CORS on R2 bucket. Needed for:
+  //   - Presigned PUT uploads from the browser (S3-style deployments).
+  //   - Browser-side fetches from JS (e.g. browser-merge in #638) on every
+  //     deployment, including Cloudflare. Native HTML elements (<video>,
+  //     <img>, <a>) load R2 cross-origin without CORS, but `fetch()` does not.
+  // Wrangler authenticates via CLOUDFLARE_API_TOKEN (or global wrangler login);
+  // R2_ACCESS_KEY_ID is for runtime signed URLs, not for setting CORS.
+  if (
+    vars.has('R2_BUCKET_NAME') &&
+    (vars.has('CLOUDFLARE_API_TOKEN') || vars.has('R2_ACCESS_KEY_ID'))
+  ) {
+    const setupCors = checkCancel(
+      await p.confirm({
+        message:
+          'Configure CORS on R2 bucket? (required for direct uploads from browser)',
+        initialValue: true,
+      })
+    );
+
+    if (setupCors) {
+      const bucketName = vars.get('R2_BUCKET_NAME');
+      if (!bucketName) {
+        p.log.error('R2_BUCKET_NAME is not set');
+        return;
+      }
+      const appUrl = vars.get('VITE_APP_URL') ?? 'http://localhost:3000';
+      const origins = new Set(['http://localhost:3000']);
+      if (appUrl !== 'http://localhost:3000') origins.add(appUrl);
+
+      const rules = buildRules({
+        origins: [...origins],
+        includeWrites: true,
+      });
+      const corsSpinner = p.spinner();
+      corsSpinner.start(`Configuring CORS on ${bucketName}`);
+      try {
+        setBucketCors(bucketName, rules);
+        corsSpinner.stop(`CORS configured on ${bucketName}`);
+      } catch {
+        corsSpinner.stop('CORS configuration failed');
+        p.log.warn(
+          'Configure manually in Cloudflare Dashboard:\n' +
+            `  R2 → ${bucketName} → Settings → CORS Policy\n\n` +
+            JSON.stringify({ rules }, null, 2)
+        );
+      }
+    }
+  }
+
+  // Offer to attach custom domains to R2 buckets
+  await attachR2Domains();
+
+  // Cloudflare Image Resizing reminder — needed for all environments since
+  // cdn-cgi/image/ URLs are served by Cloudflare's edge regardless of where the app runs
+  if (vars.has('R2_PUBLIC_STORAGE_DOMAIN')) {
+    p.log.warn(
+      chalk.bold('Cloudflare Image Resizing required\n') +
+        'Enable Image Resizing on the Cloudflare zone serving your R2 storage domain\n' +
+        `(e.g. the zone for ${vars.get('R2_PUBLIC_STORAGE_DOMAIN')}).\n` +
+        'Dashboard: Speed > Optimization > Image Resizing\n' +
+        'This is used to compress images for video generation (Kling 10MB limit).'
+    );
+  }
+
+  // -------------------------------------------------------------------------
+  // Google OAuth
+  // -------------------------------------------------------------------------
+  const googleKeys = ['GOOGLE_CLIENT_ID', 'GOOGLE_CLIENT_SECRET'] as const;
+  const hasGoogle = googleKeys.some((k) => vars.has(k));
+
+  if (hasGoogle) {
+    for (const k of googleKeys) {
+      if (vars.has(k)) p.log.success(`${k} — already configured`);
+    }
+  }
+
+  const setupGoogle =
+    hasGoogle ||
+    checkCancel(
+      await p.confirm({
+        message: 'Set up Google sign-in? (or use email OTP)',
+        initialValue: false,
+      })
+    );
+
+  if (setupGoogle) {
+    await promptForKey(
+      'GOOGLE_CLIENT_ID',
+      'Google OAuth Client ID',
+      'Get one at: https://console.cloud.google.com/apis/credentials'
+    );
+    await promptForKey('GOOGLE_CLIENT_SECRET', 'Google OAuth Client Secret');
+  }
+
+  // -------------------------------------------------------------------------
+  // Observability (Langfuse)
+  // -------------------------------------------------------------------------
+  const langfuseKeys = [
+    'LANGFUSE_PUBLIC_KEY',
+    'LANGFUSE_SECRET_KEY',
+    'LANGFUSE_BASE_URL',
+  ] as const;
+  const hasLangfuse = langfuseKeys.some((k) => vars.has(k));
+
+  if (hasLangfuse) {
+    for (const k of langfuseKeys) {
+      if (vars.has(k)) p.log.success(`${k} — already configured`);
+    }
+  }
+
+  const setupLangfuse =
+    hasLangfuse ||
+    checkCancel(
+      await p.confirm({
+        message:
+          'Set up Langfuse for prompt management & LLM tracing? (optional)',
+        initialValue: false,
+      })
+    );
+
+  if (setupLangfuse) {
+    await promptForKey(
+      'LANGFUSE_PUBLIC_KEY',
+      'Langfuse Public Key',
+      'Get one at: https://cloud.langfuse.com/project/settings'
+    );
+    await promptForKey('LANGFUSE_SECRET_KEY', 'Langfuse Secret Key');
+    await promptForKey(
+      'LANGFUSE_BASE_URL',
+      'Langfuse Base URL (leave empty for cloud)',
+      'Default: https://cloud.langfuse.com'
+    );
+
+    if (!vars.has('LANGFUSE_PROMPTS_ENABLED')) {
+      const enablePrompts = checkCancel(
+        await p.confirm({
+          message:
+            'Fetch prompts from Langfuse API? (If no, prompts are served from git)',
+          initialValue: false,
+        })
+      );
+      vars.set('LANGFUSE_PROMPTS_ENABLED', enablePrompts ? 'true' : 'false');
+      saveProgress();
+    } else {
+      p.log.success('LANGFUSE_PROMPTS_ENABLED — already configured');
+    }
+
+    const uploadPrompts = checkCancel(
+      await p.confirm({
+        message: 'Upload all prompts to Langfuse now?',
+        initialValue: true,
+      })
+    );
+
+    if (uploadPrompts) {
+      const promptSpinner = p.spinner();
+      promptSpinner.start('Uploading prompts to Langfuse');
+      try {
+        execFileSync('bun', ['scripts/upload-all-prompts.ts'], {
+          stdio: 'pipe',
+          cwd: process.cwd(),
+        });
+        promptSpinner.stop('Prompts uploaded to Langfuse');
+      } catch {
+        promptSpinner.stop('Prompt upload failed');
+        p.log.warn(
+          'You can retry later with: bun scripts/upload-all-prompts.ts'
+        );
+      }
+    }
+  }
+
+  // -------------------------------------------------------------------------
+  // Analytics (PostHog)
+  // -------------------------------------------------------------------------
+  const posthogKeys = ['VITE_PUBLIC_POSTHOG_PROJECT_TOKEN'] as const;
+  const hasPostHog = posthogKeys.some((k) => vars.has(k));
+
+  if (hasPostHog) {
+    for (const k of posthogKeys) {
+      if (vars.has(k)) p.log.success(`${k} — already configured`);
+    }
+    if (vars.has('VITE_PUBLIC_POSTHOG_HOST'))
+      p.log.success('VITE_PUBLIC_POSTHOG_HOST — already configured');
+  }
+
+  const setupPostHog =
+    hasPostHog ||
+    checkCancel(
+      await p.confirm({
+        message:
+          'Set up PostHog for product analytics & error tracking? (optional)',
+        initialValue: false,
+      })
+    );
+
+  if (setupPostHog) {
+    await promptForKey(
+      'VITE_PUBLIC_POSTHOG_PROJECT_TOKEN',
+      'PostHog Project API Key',
+      'Get one at: https://us.posthog.com/settings/project'
+    );
+    await promptForKey(
+      'VITE_PUBLIC_POSTHOG_HOST',
+      'PostHog Host (leave empty for default)',
+      'Default: https://us.i.posthog.com'
+    );
+  }
+
+  // -------------------------------------------------------------------------
+  // Billing (Stripe)
+  // -------------------------------------------------------------------------
+  const stripeKeys = ['STRIPE_SECRET_KEY', 'STRIPE_WEBHOOK_SECRET'] as const;
+  const hasStripe = stripeKeys.some((k) => vars.has(k));
+
+  if (hasStripe) {
+    for (const k of stripeKeys) {
+      if (vars.has(k)) p.log.success(`${k} — already configured`);
+    }
+  }
+
+  const setupBilling =
+    hasStripe ||
+    checkCancel(
+      await p.confirm({
+        message: 'Set up Stripe billing?',
+        initialValue: false,
+      })
+    );
+
+  if (setupBilling) {
+    await promptForKey(
+      'STRIPE_SECRET_KEY',
+      'Stripe Secret Key (sk_live_...)',
+      'Get one at: https://dashboard.stripe.com/apikeys'
+    );
+    await promptForKey(
+      'STRIPE_WEBHOOK_SECRET',
+      'Stripe Webhook Secret (whsec_...)',
+      'From your webhook endpoint in the Stripe dashboard'
+    );
+
+    const appUrl = vars.get('VITE_APP_URL') ?? '<your-app-url>';
+    p.note(
+      [
+        `Endpoint URL: ${appUrl}/api/billing/webhook`,
+        '',
+        'Events to enable:',
+        '  - checkout.session.completed',
+        '  - payment_intent.succeeded',
+        '',
+        'If using a restricted key, required permissions:',
+        '  Charges (Read), Customers (Write),',
+        '  Payment Intents (Read), Checkout Sessions (Write)',
+      ].join('\n'),
+      'Production Webhook Setup'
+    );
+  }
+
+  // -------------------------------------------------------------------------
+  // Write env file
+  // -------------------------------------------------------------------------
+  writeEnvFile(vars);
+  p.log.success(`Wrote ${ENV_FILE}`);
+
+  // -------------------------------------------------------------------------
+  // Summary
+  // -------------------------------------------------------------------------
+  const features: Array<[string, string]> = [
+    ['Hosting', 'Cloudflare Workers'],
+    ['Database', 'Cloudflare D1'],
+    ['Auth', 'Better Auth (secret generated)'],
+    [
+      'Email',
+      vars.has('EMAIL_FROM') ? 'Cloudflare Email Service' : 'Not configured',
+    ],
+    ['AI Generation', vars.has('FAL_KEY') ? 'Configured' : 'Skipped'],
+    ['Script Analysis', vars.has('OPENROUTER_KEY') ? 'Configured' : 'Skipped'],
+    [
+      'Storage',
+      vars.has('R2_PUBLIC_STORAGE_DOMAIN') ? 'Configured' : 'Skipped',
+    ],
+    ['Google OAuth', vars.has('GOOGLE_CLIENT_ID') ? 'Configured' : 'Skipped'],
+    [
+      'Langfuse',
+      vars.has('LANGFUSE_PUBLIC_KEY')
+        ? 'Configured'
+        : 'Skipped (using local prompts)',
+    ],
+    [
+      'PostHog',
+      vars.has('VITE_PUBLIC_POSTHOG_PROJECT_TOKEN') ? 'Configured' : 'Skipped',
+    ],
+    ['Stripe', vars.has('STRIPE_SECRET_KEY') ? 'Configured' : 'Not configured'],
+  ];
+
+  const summaryLines = features
+    .map(
+      ([name, status]) =>
+        `${status.includes('Skip') || status.includes('Not configured') ? '○' : '●'} ${name}: ${status}`
+    )
+    .join('\n');
+
+  p.note(summaryLines, 'Configuration Summary');
+
+  await deploySetup(vars, checkCancel);
+  p.outro(`Production config written to ${chalk.bold(ENV_FILENAME)}`);
+}
